Move UserTokenTable to account app
